chore: consolidate test-report and unbreak x-module relative imports.
