+++ packages/haiku-common/.nycrc
@@ -1,0 +1,13 @@
+{
+  "extension": [
+    ".ts"
+  ],
+  "include": [
+    "src/**/*.ts"
+  ],
+  "reporter": [
+    "cobertura",
+    "html"
+  ],
+  "all": true
+}
