+++ packages/haiku-common/package.json
@@ -12,8 +12,7 @@
     "fix": "yarn lint --fix",
     "ts": "cross-env NODE_ENV=test yarn ts-node -r tsconfig-paths/register -P tsconfig.all.json",
     "test": "yarn ts ./node_modules/.bin/tape 'test/**/*.test.ts' | tap-spec",
-    "test-report": "yarn ts ./node_modules/.bin/tape 'test/**/*.test.ts' > test-result.tap",
-    "coverage": "nyc yarn test",
+    "test-report": "nyc yarn ts ./node_modules/.bin/tape 'test/**/*.test.ts' > test-result.tap",
     "compile": "tsc",
     "develop": "tsc --watch"
   },
